+++ src/ManipulatorPlugin/MprPositionListView.cpp
@@ -749,7 +749,7 @@ MprPositionListView::Impl::Impl(MprPositionListView* self)
 
     // Setup the table
     auto hframe = new QFrame;
-    hframe->setFrameStyle(QFrame::HLine | QFrame::Sunken);
+    hframe->setFrameStyle(static_cast<int>(QFrame::HLine) | static_cast<int>(QFrame::Sunken));
     vbox->addWidget(hframe);
     setFrameShape(QFrame::NoFrame);
     setSelectionBehavior(QAbstractItemView::SelectRows);
